Switch to using time()
clock() had some weird issues on Burrow - it looks like it was returning tenths of a second. Changed to using time() which has decent enough resolution for what we're using it for (Depth=3 dies with half a second left with a 5 second timer, etc.). Changed how we print out: if depth=3, we always print. If depth > 3, we only print if the timer has not yet expired (meaning the search completed fully).